Module rust_gpu_sdf::operators

Types that modify a distance field.

Modules

axial_reflect

Reflect a distance field about an arbitrary axis.

compose

composite

Operators composed from other operators.

conditional

displace

Displace the output of a distance field using the output of another distance field.

elongate

Extrude a shape along its axes, preserving exterior geometry.

hollow

Convert a solid shape into a hollow one with an infinitely thin surface.

intersection

Compute the boolean intersection of two distance fields.

isosurface

Shift the isosurface of a distance field by a given amount.

reflect

Reflect a distance field about an arbitrary axis.

repeat

Operators for repeating distance fields across a domain.

rotate

Rotate a distance field.

scale

Uniformly scale a distance field.

sided

Given an infinitely-thin surface, divide space into interior and exterior based on axis.

smooth_intersection

Compute the blended boolean intersection of two distance fields.

smooth_subtraction

Compute the blended boolean subtraction of two distance fields.

smooth_union

Compute the blended boolean union of two distance fields.

stretch

Stretch a shape along an arbitrary axis, preserving exterior geometry as caps.

subtraction

Compute the boolean subtraction of two distance fields.

translate

Apply a positional translation to a distance field.

twist

Twist a distance field around an arbitrary axis.

union

Compute the boolean union of two distance fields.

Structs

Operator

Applies a SignedDistanceOperator to a SignedDistanceField.

Traits

SignedDistanceOperator

Modifies the input / output of a SignedDistanceField.